+++ src/qgis_geonode/apiclient/base.py
@@ -4,11 +4,14 @@ from functools import partial
 import qgis.core
 from qgis.PyQt import (
     QtCore,
-    QtNetwork,
     QtXml,
 )
 
-from .. import network, conf
+from .. import (
+    conf,
+    network,
+)
+
 from . import models
 from .models import GeonodeApiSearchFilters
 
@@ -25,8 +28,11 @@ class BaseGeonodeClient(QtCore.QObject):
     dataset_detail_received = QtCore.pyqtSignal(object)
     dataset_detail_error_received = QtCore.pyqtSignal([str], [str, int, str])
     style_detail_received = QtCore.pyqtSignal(QtXml.QDomElement)
+    style_detail_error_received = QtCore.pyqtSignal([str], [str, int, str])
     keyword_list_received = QtCore.pyqtSignal(list)
     search_error_received = QtCore.pyqtSignal([str], [str, int, str])
+    dataset_uploaded = QtCore.pyqtSignal(int)
+    dataset_upload_error_received = QtCore.pyqtSignal([str], [str, int, str])
 
     def __init__(
         self,
@@ -65,11 +71,14 @@ class BaseGeonodeClient(QtCore.QObject):
     def get_dataset_detail_url(self, dataset_id: int) -> QtCore.QUrl:
         raise NotImplementedError
 
-    def get_dataset_list(self, search_filters: GeonodeApiSearchFilters):
+    def get_dataset_upload_url(self) -> QtCore.QUrl:
+        raise NotImplementedError
+
+    def get_dataset_list(self, search_filters: GeonodeApiSearchFilters) -> None:
         self.network_fetcher_task = network.NetworkRequestTask(
             [network.RequestToPerform(url=self.get_dataset_list_url(search_filters))],
+            self.network_requests_timeout,
             self.auth_config,
-            network_task_timeout=self.network_requests_timeout,
             description="Get dataset list",
         )
         self.network_fetcher_task.task_done.connect(self.handle_dataset_list)
@@ -82,7 +91,33 @@ class BaseGeonodeClient(QtCore.QObject):
         """
         raise NotImplementedError
 
-    def get_dataset_detail(self, brief_dataset: models.BriefDataset):
+    def get_dataset_style(
+        self, dataset: models.Dataset, emit_dataset_detail_received: bool = False
+    ) -> None:
+        self.network_fetcher_task = network.NetworkRequestTask(
+            [network.RequestToPerform(QtCore.QUrl(dataset.default_style.sld_url))],
+            self.network_requests_timeout,
+            self.auth_config,
+            description="Get dataset style",
+        )
+        self.network_fetcher_task.task_done.connect(
+            partial(
+                self.handle_dataset_style,
+                dataset,
+                emit_dataset_detail_received=emit_dataset_detail_received,
+            )
+        )
+        qgis.core.QgsApplication.taskManager().addTask(self.network_fetcher_task)
+
+    def handle_dataset_style(
+        self,
+        dataset: models.Dataset,
+        task_result: bool,
+        emit_dataset_detail_received: typing.Optional[bool] = False,
+    ) -> None:
+        raise NotImplementedError
+
+    def get_dataset_detail(self, brief_dataset: models.BriefDataset) -> None:
         requests_to_perform = [
             network.RequestToPerform(url=self.get_dataset_detail_url(brief_dataset.pk))
         ]
@@ -92,8 +127,8 @@ class BaseGeonodeClient(QtCore.QObject):
 
         self.network_fetcher_task = network.NetworkRequestTask(
             requests_to_perform,
+            self.network_requests_timeout,
             self.auth_config,
-            network_task_timeout=self.network_requests_timeout,
             description="Get dataset detail",
         )
         self.network_fetcher_task.task_done.connect(
@@ -102,4 +137,48 @@ class BaseGeonodeClient(QtCore.QObject):
         qgis.core.QgsApplication.taskManager().addTask(self.network_fetcher_task)
 
     def handle_dataset_detail(self, brief_dataset: models.BriefDataset, result: bool):
+        """Handle dataset detail retrieval outcome.
+
+        This method should emit either `dataset_detail_received` or
+        `dataset_detail_error_received`.
+
+        """
+
+        raise NotImplementedError
+
+    def get_dataset_detail_from_id(self, dataset_id: int):
+        self.network_fetcher_task = network.NetworkRequestTask(
+            [network.RequestToPerform(url=self.get_dataset_detail_url(dataset_id))],
+            self.network_requests_timeout,
+            self.auth_config,
+            description="Get dataset detail",
+        )
+        self.network_fetcher_task.task_done.connect(self.handle_dataset_detail_from_id)
+        qgis.core.QgsApplication.taskManager().addTask(self.network_fetcher_task)
+
+    def handle_dataset_detail_from_id(self, task_result: bool):
+        raise NotImplementedError
+
+    def get_uploader_task(
+        self, layer: qgis.core.QgsMapLayer, allow_public_access: bool, timeout: int
+    ) -> qgis.core.QgsTask:
+        raise NotImplementedError
+
+    def upload_layer(
+        self, layer: qgis.core.QgsMapLayer, allow_public_access: bool
+    ) -> None:
+        self.network_fetcher_task = self.get_uploader_task(
+            layer, allow_public_access, timeout=10 * 60 * 1000
+        )  # the GeoNode GUI also uses a 10 minute timeout for uploads
+        self.network_fetcher_task.task_done.connect(self.handle_layer_upload)
+        qgis.core.QgsApplication.taskManager().addTask(self.network_fetcher_task)
+
+    def handle_layer_upload(self, result: bool):
+        """Handle layer upload outcome.
+
+        This method should emit either `dataset_uploaded` or
+        `dataset_upload_error_received`.
+
+        """
+
         raise NotImplementedError