+++ src/qgis_geonode/apiclient/version_postv2.py
@@ -1,13 +1,24 @@
+import dataclasses
 import datetime as dt
+import json
+import shutil
+import tempfile
 import typing
 import uuid
+from pathlib import Path
 
 import qgis.core
-from qgis.PyQt import QtCore
+import qgis.utils
+from qgis.PyQt import (
+    QtCore,
+    QtNetwork,
+)
 
 from .. import network
 from .. import styles as geonode_styles
-from ..utils import log
+from ..utils import (
+    log,
+)
 
 from . import models
 from .base import BaseGeonodeClient
@@ -35,6 +46,8 @@ class GeonodePostV2ApiClient(BaseGeonodeClient):
         models.ApiClientCapability.LOAD_VECTOR_DATASET_VIA_WFS,
         models.ApiClientCapability.LOAD_RASTER_DATASET_VIA_WMS,
         models.ApiClientCapability.LOAD_RASTER_DATASET_VIA_WCS,
+        models.ApiClientCapability.UPLOAD_VECTOR_LAYER,
+        models.ApiClientCapability.UPLOAD_RASTER_LAYER,
     ]
 
     @property
@@ -129,112 +142,440 @@ class GeonodePostV2ApiClient(BaseGeonodeClient):
     def get_layer_style_list_url(self, layer_id: int):
         return QtCore.QUrl(f"{self.dataset_list_url}{layer_id}/styles/")
 
-    def handle_dataset_list(self, result: bool) -> None:
-        log(f"inside apiclient.handle_dataset_list with a result of {result!r}")
-        if result:
-            response_content: network.ParsedNetworkReply = (
-                self.network_fetcher_task.response_contents[0]
+    def get_dataset_upload_url(self) -> QtCore.QUrl:
+        return QtCore.QUrl(f"{self.api_url}/uploads/upload/")
+
+    def handle_dataset_list(self, task_result: bool) -> None:
+        deserialized_content = self._retrieve_response(
+            task_result, 0, self.search_error_received
+        )
+        if deserialized_content is not None:
+            brief_datasets = []
+            for raw_brief_dataset in deserialized_content.get("datasets", []):
+                try:
+                    parsed_properties = _get_common_model_properties(raw_brief_dataset)
+                    brief_dataset = models.BriefDataset(**parsed_properties)
+                except ValueError:
+                    log(
+                        f"Could not parse {raw_brief_dataset!r} into " f"a valid item",
+                        debug=False,
+                    )
+                else:
+                    brief_datasets.append(brief_dataset)
+            pagination_info = models.GeonodePaginationInfo(
+                total_records=deserialized_content.get("total") or 0,
+                current_page=deserialized_content.get("page") or 1,
+                page_size=deserialized_content.get("page_size") or 0,
             )
+            self.dataset_list_received.emit(brief_datasets, pagination_info)
+
+    def handle_dataset_detail(
+        self, brief_dataset: models.BriefDataset, task_result: bool
+    ) -> None:
+        deserialized_resource = self._retrieve_response(
+            task_result, 0, self.dataset_detail_error_received
+        )
+        if deserialized_resource is not None:
+            try:
+                dataset = parse_dataset_detail(deserialized_resource["dataset"])
+            except KeyError as exc:
+                log(
+                    f"Could not parse server response into a dataset: {str(exc)}",
+                    debug=False,
+                )
+            else:
+                is_vector = (
+                    brief_dataset.dataset_sub_type
+                    == models.GeonodeResourceType.VECTOR_LAYER
+                )
+                if is_vector:
+                    (sld_named_layer, error_message,) = geonode_styles.get_usable_sld(
+                        self.network_fetcher_task.response_contents[1]
+                    )
+                    if sld_named_layer is None:
+                        raise RuntimeError(error_message)
+                    dataset.default_style.sld = sld_named_layer
+                self.dataset_detail_received.emit(dataset)
+
+    def handle_dataset_detail_from_id(self, task_result: bool) -> None:
+        deserialized_resource = self._retrieve_response(
+            task_result, 0, self.dataset_detail_error_received
+        )
+        if deserialized_resource is not None:
+            try:
+                dataset = parse_dataset_detail(deserialized_resource["dataset"])
+            except KeyError as exc:
+                log(
+                    f"Could not parse server response into a dataset: {str(exc)}",
+                    debug=False,
+                )
+            else:
+                if dataset.dataset_sub_type == models.GeonodeResourceType.VECTOR_LAYER:
+                    self.get_dataset_style(dataset, emit_dataset_detail_received=True)
+                else:
+                    self.dataset_detail_received.emit(dataset)
+
+    def handle_dataset_style(
+        self,
+        dataset: models.Dataset,
+        task_result: bool,
+        emit_dataset_detail_received: bool = False,
+    ) -> None:
+        response_contents = self._retrieve_response(
+            task_result, 0, self.style_detail_error_received, deserialize_as_json=False
+        )
+        if response_contents is not None:
+            sld_named_layer, error_message = geonode_styles.get_usable_sld(
+                response_contents
+            )
+            if sld_named_layer is None:
+                self.style_detail_error_received[str].emit(
+                    f"Could not parse downloaded SLD: {error_message}"
+                )
+            dataset.default_style.sld = sld_named_layer
+            if emit_dataset_detail_received:
+                self.dataset_detail_received.emit(dataset)
+
+    def get_uploader_task(
+        self, layer: qgis.core.QgsMapLayer, allow_public_access: bool, timeout: int
+    ) -> qgis.core.QgsTask:
+        return LayerUploaderTask(
+            layer,
+            self.get_dataset_upload_url(),
+            allow_public_access,
+            self.auth_config,
+            network_task_timeout=timeout,
+            description="Upload layer to GeoNode",
+        )
+
+    def handle_layer_upload(self, result: bool):
+        if result:
+            response_contents = self.network_fetcher_task.response_contents[0]
+            if response_contents.http_status_code == 201:
+                deserialized = network.deserialize_json_response(
+                    response_contents.response_body
+                )
+                catalogue_url = deserialized["url"]
+                dataset_pk = catalogue_url.rsplit("/")[-1]
+                self.dataset_uploaded.emit(int(dataset_pk))
+            else:
+                self.dataset_upload_error_received[str, int, str].emit(
+                    response_contents.qt_error,
+                    response_contents.http_status_code,
+                    response_contents.http_status_reason,
+                )
+        else:
+            self.dataset_upload_error_received[str].emit(
+                "Could not upload layer to GeoNode"
+            )
+
+    def _retrieve_response(
+        self,
+        task_result: bool,
+        contents_index: int,
+        error_signal,
+        deserialize_as_json: typing.Optional[bool] = True,
+    ) -> typing.Optional[typing.Union[typing.Dict, network.ParsedNetworkReply]]:
+        """Internal method that takes care of boilerplate-ish response parsing."""
+        result = None
+        if task_result:
+            response_content = self.network_fetcher_task.response_contents[
+                contents_index
+            ]
             if response_content.qt_error is None:
-                if not response_content.response_body.isEmpty():
-                    deserialized_content = network.deserialize_json_response(
+                result = response_content
+                if deserialize_as_json:
+                    deserialized = network.deserialize_json_response(
                         response_content.response_body
                     )
-                    if deserialized_content is not None:
-                        brief_datasets = []
-                        for raw_brief_dataset in deserialized_content.get(
-                            "datasets", []
-                        ):
-                            try:
-                                parsed_properties = _get_common_model_properties(
-                                    raw_brief_dataset
-                                )
-                                brief_dataset = models.BriefDataset(**parsed_properties)
-                            except ValueError:
-                                log(
-                                    f"Could not parse {raw_brief_dataset!r} into "
-                                    f"a valid item",
-                                    debug=False,
-                                )
-                            else:
-                                brief_datasets.append(brief_dataset)
-                        pagination_info = models.GeonodePaginationInfo(
-                            total_records=deserialized_content.get("total") or 0,
-                            current_page=deserialized_content.get("page") or 1,
-                            page_size=deserialized_content.get("page_size") or 0,
-                        )
-                        self.dataset_list_received.emit(brief_datasets, pagination_info)
+                    if deserialized is not None:
+                        result = deserialized
                     else:
-                        self.search_error_received[str].emit(
-                            "Could not parse dataset list returned from remote GeoNode"
+                        error_signal[str].emit(
+                            "Could not parse response from remote GeoNode"
                         )
-                else:
-                    self.search_error_received[str, int, str].emit(
-                        response_content.qt_error,
-                        response_content.http_status_code,
-                        response_content.http_status_reason,
-                    )
             else:
-                self.search_error_received[str, int, str].emit(
+                error_signal[str, int, str].emit(
                     response_content.qt_error,
                     response_content.http_status_code,
                     response_content.http_status_reason,
                 )
         else:
-            self.search_error_received[str].emit(
-                "Could not complete request for dataset list"
-            )
+            error_signal[str].emit("Could not complete network request")
+        return result
 
-    def handle_dataset_detail(
-        self, brief_dataset: models.BriefDataset, result: bool
-    ) -> None:
-        if result:
-            detail_response_content: network.ParsedNetworkReply = (
-                self.network_fetcher_task.response_contents[0]
+
+@dataclasses.dataclass()
+class ExportFormat:
+    driver_name: str
+    file_extension: str
+
+
+class LayerUploaderTask(network.NetworkRequestTask):
+    VECTOR_UPLOAD_FORMAT: typing.Final[ExportFormat] = ExportFormat(
+        "ESRI Shapefile", "shp"
+    )
+    RASTER_UPLOAD_FORMAT: typing.Final[ExportFormat] = ExportFormat("GTiff", "tif")
+
+    layer: qgis.core.QgsMapLayer
+    allow_public_access: bool
+    _upload_url: QtCore.QUrl
+    _temporary_directory: typing.Optional[Path]
+
+    def __init__(
+        self,
+        layer: qgis.core.QgsMapLayer,
+        upload_url: QtCore.QUrl,
+        allow_public_access: bool,
+        authcfg: str,
+        network_task_timeout: int,
+        description: str = "LayerUploaderTask",
+    ):
+        """Task to perform upload of QGIS layers to remote GeoNode servers."""
+        super().__init__(
+            requests_to_perform=[],
+            authcfg=authcfg,
+            description=description,
+            network_task_timeout=network_task_timeout,
+        )
+        self.response_contents = [None]
+        self.layer = layer
+        self.allow_public_access = allow_public_access
+        self._upload_url = upload_url
+        self._temporary_directory = None
+
+    def run(self) -> bool:
+        if self._is_layer_uploadable():
+            source_path = Path(
+                self.layer.dataProvider().dataSourceUri().partition("|")[0]
             )
-            empty_body = detail_response_content.response_body.isEmpty()
-            if detail_response_content.qt_error is None and not empty_body:
-                deserialized_resource = network.deserialize_json_response(
-                    detail_response_content.response_body
-                )
-                if deserialized_resource is not None:
-                    try:
-                        dataset = parse_dataset_detail(deserialized_resource["dataset"])
-                    except KeyError as exc:
-                        log(
-                            f"Could not parse server response into a dataset: {str(exc)}",
-                            debug=False,
-                        )
-                    else:
-                        is_vector = (
-                            brief_dataset.dataset_sub_type
-                            == models.GeonodeResourceType.VECTOR_LAYER
-                        )
-                        if is_vector:
-                            (
-                                sld_named_layer,
-                                error_message,
-                            ) = geonode_styles.get_usable_sld(
-                                self.network_fetcher_task.response_contents[1]
-                            )
-                            if sld_named_layer is None:
-                                raise RuntimeError(error_message)
-                            dataset.default_style.sld = sld_named_layer
-                        self.dataset_detail_received.emit(dataset)
-                else:
-                    self.dataset_detail_error_received[str].emit(
-                        "Could not parse dataset detail returned from remote GeoNode"
-                    )
-            else:
-                self.dataset_detail_error_received[str, int, str].emit(
-                    detail_response_content.qt_error,
-                    detail_response_content.http_status_code,
-                    detail_response_content.http_status_reason,
-                )
+            export_error = None
         else:
-            self.dataset_detail_error_received[str].emit(
-                "Could not complete request for dataset detail"
+            log(
+                "Exporting layer to an uploadable format before proceeding with "
+                "the upload..."
             )
+            source_path, export_error = self._export_layer_to_temp_dir()
+        log(f"source_path: {source_path}")
+        if export_error is None:
+            multipart = self._prepare_multipart(source_path)
+            with network.wait_for_signal(
+                self._all_requests_finished, timeout=self.network_task_timeout
+            ) as event_loop_result:
+                request = QtNetwork.QNetworkRequest(self._upload_url)
+                request.setHeader(
+                    QtNetwork.QNetworkRequest.ContentTypeHeader,
+                    f"multipart/form-data; boundary={multipart.boundary().data().decode()}",
+                )
+                if self.authcfg:
+                    auth_manager = qgis.core.QgsApplication.authManager()
+                    auth_added, _ = auth_manager.updateNetworkRequest(
+                        request, self.authcfg
+                    )
+                else:
+                    auth_added = True
+                if auth_added:
+                    qt_reply = self._dispatch_request(
+                        request, network.HttpMethod.POST, multipart
+                    )
+                    multipart.setParent(qt_reply)
+                    request_id = qt_reply.property("requestId")
+                    self._pending_replies[request_id] = (0, qt_reply)
+                else:
+                    self._all_requests_finished.emit()
+            loop_forcibly_ended = not bool(event_loop_result.result)
+            if loop_forcibly_ended:
+                result = False
+            else:
+                result = self._num_finished >= len(self.requests_to_perform)
+        else:
+            result = False
+        return result
+
+    def finished(self, result: bool) -> None:
+        if self._temporary_directory is not None:
+            shutil.rmtree(self._temporary_directory, ignore_errors=True)
+        super().finished(result)
+
+    def _prepare_multipart(self, source_path: Path) -> QtNetwork.QHttpMultiPart:
+        main_file = QtCore.QFile(str(source_path))
+        main_file.open(QtCore.QIODevice.ReadOnly)
+        sidecar_files = []
+        if self.layer.type() == qgis.core.QgsMapLayerType.VectorLayer:
+            dbf_file = QtCore.QFile(str(source_path.parent / f"{source_path.stem}.dbf"))
+            dbf_file.open(QtCore.QIODevice.ReadOnly)
+            sidecar_files.append(("dbf_file", dbf_file))
+            prj_file = QtCore.QFile(str(source_path.parent / f"{source_path.stem}.prj"))
+            prj_file.open(QtCore.QIODevice.ReadOnly)
+            sidecar_files.append(("prj_file", prj_file))
+            shx_file = QtCore.QFile(str(source_path.parent / f"{source_path.stem}.shx"))
+            shx_file.open(QtCore.QIODevice.ReadOnly)
+            sidecar_files.append(("shx_file", shx_file))
+        elif self.layer.type() == qgis.core.QgsMapLayerType.RasterLayer:
+            # when uploading tif files GeoNode seems to want the same file be uploaded
+            # twice - one under the `base_file` form field and another under the
+            # `tif_file` form field. This seems like a bug in GeoNode though
+            tif_file = QtCore.QFile(str(source_path))
+            tif_file.open(QtCore.QIODevice.ReadOnly)
+            sidecar_files.append(("tif_file", tif_file))
+        permissions = {
+            "users": {},
+            "groups": {},
+        }
+        if self.allow_public_access:
+            permissions["users"]["AnonymousUser"] = [
+                "view_resourcebase",
+                "download_resourcebase",
+            ]
+        multipart = build_multipart(
+            self.layer.metadata(), permissions, main_file, sidecar_files=sidecar_files
+        )
+        # below we set all QFiles as children of the multipart object and later we
+        # also make the multipart object a children on the network reply object. This is
+        # done in order to ensure deletion of resources at the correct time, as
+        # recommended by the Qt documentation at:
+        # https://doc.qt.io/qt-5/qhttppart.html#details
+        main_file.setParent(multipart)
+        for _, qt_file in sidecar_files:
+            qt_file.setParent(multipart)
+        return multipart
+
+    def _is_layer_uploadable(self) -> bool:
+        """Check if the layer is in a format suitable for uploading to GeoNode."""
+        ds_uri = self.layer.dataProvider().dataSourceUri()
+        fragment = ds_uri.split("|")[0]
+        extension = fragment.rpartition(".")[-1]
+        return extension in (
+            self.VECTOR_UPLOAD_FORMAT.file_extension,
+            self.RASTER_UPLOAD_FORMAT.file_extension,
+        )
+
+    def _export_layer_to_temp_dir(
+        self,
+    ) -> typing.Tuple[typing.Optional[Path], typing.Optional[str]]:
+        self._temporary_directory = Path(tempfile.mkdtemp(prefix="qgis_geonode_"))
+        error_message = None
+        if self.layer.type() == qgis.core.QgsMapLayerType.VectorLayer:
+            exported_path, error_message = self._export_vector_layer()
+        elif self.layer.type() == qgis.core.QgsMapLayerType.RasterLayer:
+            exported_path, export_error = self._export_raster_layer()
+        else:
+            raise NotImplementedError()
+        return exported_path, error_message or None
+
+    def _export_vector_layer(
+        self,
+    ) -> typing.Tuple[typing.Optional[Path], str]:
+        sanitized_layer_name = network.sanitize_layer_name(self.layer.name())
+        target_path = self._temporary_directory / f"{sanitized_layer_name}.shp"
+        export_code, error_message = qgis.core.QgsVectorLayerExporter.exportLayer(
+            layer=self.layer,
+            uri=str(target_path),
+            providerKey="ogr",
+            destCRS=qgis.core.QgsCoordinateReferenceSystem(),
+            options={
+                "driverName": "ESRI Shapefile",
+            },
+        )
+        if export_code == qgis.core.Qgis.VectorExportResult.Success:
+            result = (target_path, error_message)
+        else:
+            result = (None, error_message)
+        return result
+
+    def _export_raster_layer(
+        self,
+    ) -> typing.Tuple[typing.Optional[Path], typing.Optional[int]]:
+        sanitized_layer_name = network.sanitize_layer_name(self.layer.name())
+        target_path = (
+            self._temporary_directory
+            / f"{sanitized_layer_name}.{self.RASTER_UPLOAD_FORMAT.file_extension}"
+        )
+        writer = qgis.core.QgsRasterFileWriter(str(target_path))
+        writer.setOutputFormat(self.RASTER_UPLOAD_FORMAT.driver_name)
+        pipe = self.layer.pipe()
+        raster_interface = self.layer.dataProvider()
+        write_error = writer.writeRaster(
+            pipe,
+            raster_interface.xSize(),
+            raster_interface.ySize(),
+            raster_interface.extent(),
+            raster_interface.crs(),
+            qgis.core.QgsCoordinateTransformContext(),
+        )
+        if write_error == qgis.core.QgsRasterFileWriter.NoError:
+            result = (target_path, None)
+        else:
+            result = (None, write_error)
+        return result
+
+
+def build_multipart(
+    layer_metadata: qgis.core.QgsLayerMetadata,
+    permissions: typing.Dict,
+    main_file: QtCore.QFile,
+    sidecar_files: typing.List[typing.Tuple[str, QtCore.QFile]],
+) -> QtNetwork.QHttpMultiPart:
+    encoding = "utf-8"
+    multipart = QtNetwork.QHttpMultiPart(QtNetwork.QHttpMultiPart.FormDataType)
+    title = layer_metadata.title()
+    if title:
+        title_part = QtNetwork.QHttpPart()
+        title_part.setHeader(
+            QtNetwork.QNetworkRequest.ContentDispositionHeader,
+            'form-data; name="dataset_title"',
+        )
+        title_part.setBody(layer_metadata.title().encode(encoding))
+        multipart.append(title_part)
+    abstract = layer_metadata.abstract()
+    if abstract:
+        abstract_part = QtNetwork.QHttpPart()
+        abstract_part.setHeader(
+            QtNetwork.QNetworkRequest.ContentDispositionHeader,
+            'form-data; name="abstract"',
+        )
+        abstract_part.setBody(layer_metadata.abstract().encode(encoding))
+        multipart.append(abstract_part)
+    false_items = (
+        "time",
+        "mosaic",
+        "metadata_uploaded_preserve",
+        "metadata_upload_form",
+        "style_upload_form",
+    )
+    for item in false_items:
+        part = QtNetwork.QHttpPart()
+        part.setHeader(
+            QtNetwork.QNetworkRequest.ContentDispositionHeader,
+            f'form-data; name="{item}"',
+        )
+        part.setBody("false".encode("utf-8"))
+        multipart.append(part)
+    permissions_part = QtNetwork.QHttpPart()
+    permissions_part.setHeader(
+        QtNetwork.QNetworkRequest.ContentDispositionHeader,
+        'form-data; name="permissions"',
+    )
+    permissions_part.setBody(json.dumps(permissions).encode(encoding))
+    multipart.append(permissions_part)
+    file_parts = [("base_file", main_file)]
+    for additional_file_form_name, additional_file_handler in sidecar_files:
+        file_parts.append((additional_file_form_name, additional_file_handler))
+    for form_element_name, file_handler in file_parts:
+        file_name = file_handler.fileName().rpartition("/")[-1]
+        part = QtNetwork.QHttpPart()
+        part.setHeader(
+            QtNetwork.QNetworkRequest.ContentDispositionHeader,
+            f'form-data; name="{form_element_name}"; filename="{file_name}"',
+        )
+        if file_name.rpartition(".")[-1] == "tif":
+            content_type = "image/tiff"
+        else:
+            content_type = "application/qgis"
+        part.setHeader(QtNetwork.QNetworkRequest.ContentTypeHeader, content_type)
+        part.setBodyDevice(file_handler)
+        multipart.append(part)
+    return multipart
 
 
 def parse_dataset_detail(raw_dataset: typing.Dict) -> models.Dataset: