+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Allow uploading QGIS layers to GeoNode as new datasets
+
 
 ## [0.4.0] - 2021-12-20